refactor: habitsTable,emoji 컴포넌트 분리
