+++ src/App.jsx
@@ -1,15 +1,12 @@
 import { TodayFocus } from './pages/TodayFocus.jsx';
 import { UiPreview } from '@/pages/UiPreview.jsx';
 import { CommonToaster } from './components/toasterConfig.jsx';
-// import { AppRouter } from './routes/AppRouter.jsx';
+import { AppRouter } from './routes/AppRouter.jsx';
 
 function App() {
   return (
     <>
-      {/* <AppRouter /> */}
-      <TodayFocus />
-      <UiPreview />
-      <CommonToaster />
+      <AppRouter />
     </>
   );
 }
